refactor(Text Column): pull out code for text column banner into a reusable component.
